Split doc/{COPYING => COPYING, COPYING.LGPL}
Split into separate files for the GPL and LGPL to make things slightly
more obvious.  Also updates the surface mail address according to the
FSF's latest template, and fixes a left-over reference to the ``GNU Library
General Public License'', which is now known as the ``GNU Lesser General
Public License''.