setfacl: changing owner and when S_ISUID should be set --restore fix
Fix a problem in setfacl --restore when the owner or group is changed
and the S_ISUID and S_ISGID are to be set.

The root of the problem is that chown() can clear the S_ISUID and
S_ISGID bits as described in chown(2):

 When  the  owner  or  group of an executable file are changed by a
 non- superuser, the S_ISUID and S_ISGID mode bits are cleared.   POSIX
 does not specify whether this also should happen when root does the
 chown(); the Linux behavior depends on the kernel version.  In case  of
 a  non- group-executable  file (i.e., one for which the S_IXGRP bit is
 not set) the S_ISGID bit indicates mandatory locking, and is not
 cleared  by  a chown().

To fix the issue re-stat() the file after chown() so that the logic
surrounding the chmod() has the updated mode of the file.

Signed-off-by: Brandon Philips <bphilips@suse.de>